orbfit: Redirect all output to $ORBFIT_LOG file
If ORBFIT_LOG environment variable is set, write all output that used to go
to stderr to that file. Otherwise, write it to /dev/null.

This is implemented by opening $ORBFIT_LOG (or /dev/null) as unit 99, and
writing everything that went to unit 0 there (which is why so many files are
touched by this commit).