calcpos.x: Efficient multi-epoch ephem. computation mode
Added a fourth invocation variant:

  calcpos.x <epochs.txt> <catalog_id> <ASTORB2|NATIVE|COMET> [ast_first ast_last]

where a file with a list of MJDs can be given for which to compute the
positions of a block of asteroids from the input catalog.  Argument
<epochs.txt> is either a filename, or '-' in which case the contents of
epochs.txt will be read from the standard input.  The epoch.txt file format
is as follows:

	t0  ra1 dec1 radius1
	t1  ra2 dec2 radius2
	t3  ra3 dec3 radius3
	....

where t is the time (in MJD) at which to compute the ephemeris. Optional
parameters ra, dec and radius will cause only the asteroids that are within
'radius' degrees of (ra, dec) at time 't' to be printed out. Example:

	51032.2233	11.2123 23.2233  2.5
	52032.2233	17.2123 24.2233  1.5
	50332.2233
	51332.2233	17.2123 24.2233  1.5

Arguments ast_first and ast_last specify the range of asteroids from the
input catalog for which to compute the ephemeris.  They can be given as
percentages (e.g., '10%').  Example:

	calcpos.x epochs.txt latest ASTORB2 10% 20%

will have calcpos operate on asteroids between 0.1*NTOTAL and 0.2*NTOTAL,
where NTOTAL is the number of objects in the input catalog.